Migrate terminal geometry to Swift Int

+++ Examples/DiffScope/Tests/DiffScopeExampleCoreTests/DiffScopeExampleTests.swift
@@ -200,7 +200,7 @@ private actor DiffLoadCounter {
     #expect(rangePatch.contains { $0.text == "+let value = 2" })
   }
 
-  private func render<W: Widget>(_ screen: W, width: UInt16, height: UInt16) -> String {
+  private func render<W: Widget>(_ screen: W, width: Int, height: Int) -> String {
     let area = Rect(x: 0, y: 0, width: width, height: height)
     var buffer = Buffer(area: area)
     screen.render(in: area, into: &buffer, environment: RenderEnvironment())
